backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
email-validator==2.0.0
python-dotenv==1.0.0
